Timer, pintado mejorado y prueba rapida

+++ XML/XMLParser.hpp
@@ -147,7 +147,7 @@ public:
     {
         countries.find_child_by_attribute("id", "paisesBlancosVal").text().data().set_value(to_string(cantidad).c_str());
     }
-    void cambiarTiempo(int segundos)
+    void cambiarTiempo(float segundos)
     {
         string mensaje = to_string(segundos) + " segundos";
         countries.find_child_by_attribute("id", "tiempoVal").text().data().set_value(mensaje.c_str());
